fix(l10n): restructure pt_BR contribution as base pt locale
The PR #183 contribution hand-edited the generated localization files
because gen-l10n rejects a country-coded arb (app_pt_BR.arb) without a
base app_pt.arb. The hand-edit missed the delegate isSupported list and
the lookupAppLocalizations switch, so picking the language fell back to
the first supported locale instead of Portuguese, and the committed arb
had a trailing comma that broke gen-l10n entirely.

Reshape to a plain base locale like the existing eight languages:
- rename app_pt_BR.arb -> app_pt.arb, @@locale pt, fix trailing comma
- picker code pt_BR -> pt (label stays Portugues (Brasil))
- regenerate via flutter gen-l10n (replaces the hand-edited files)

Translation content is untouched: all 543 keys, verified complete with
matching placeholders. flutter analyze passes.
