+++ mobile/lib/l10n/app_localizations.dart
@@ -11,9 +11,9 @@ import 'app_localizations_es.dart';
 import 'app_localizations_fr.dart';
 import 'app_localizations_it.dart';
 import 'app_localizations_pl.dart';
+import 'app_localizations_pt.dart';
 import 'app_localizations_ru.dart';
 import 'app_localizations_zh.dart';
-import 'app_localizations_pt_BR.dart';
 
 // ignore_for_file: type=lint
 
@@ -107,9 +107,9 @@ abstract class AppLocalizations {
     Locale('fr'),
     Locale('it'),
     Locale('pl'),
+    Locale('pt'),
     Locale('ru'),
-    Locale('zh'),
-    Locale('pt', 'BR')
+    Locale('zh')
   ];
 
   /// Title of the in-app update dialog while the new APK is downloading.
@@ -3388,6 +3388,7 @@ class _AppLocalizationsDelegate
         'fr',
         'it',
         'pl',
+        'pt',
         'ru',
         'zh'
       ].contains(locale.languageCode);
@@ -3411,6 +3412,8 @@ AppLocalizations lookupAppLocalizations(Locale locale) {
       return AppLocalizationsIt();
     case 'pl':
       return AppLocalizationsPl();
+    case 'pt':
+      return AppLocalizationsPt();
     case 'ru':
       return AppLocalizationsRu();
     case 'zh':
